# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ElevateKe Investment Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/users/{id}:
    get:
      operationId: getAdminUser
      tags: [admin]
      summary: Get user detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserDetail"
  /admin/users/{id}/suspend:
    patch:
      operationId: suspendUser
      tags: [admin]
      summary: Suspend a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User suspended
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/users/{id}/unsuspend:
    patch:
      operationId: unsuspendUser
      tags: [admin]
      summary: Unsuspend a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User unsuspended
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/plans:
    get:
      operationId: getAdminPlans
      tags: [admin]
      summary: Get all deposit plans (admin)
      responses:
        "200":
          description: Plans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DepositPlan"
    post:
      operationId: createPlan
      tags: [admin]
      summary: Create a deposit plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePlanBody"
      responses:
        "201":
          description: Plan created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DepositPlan"
  /admin/plans/{id}:
    patch:
      operationId: updatePlan
      tags: [admin]
      summary: Update a deposit plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePlanBody"
      responses:
        "200":
          description: Plan updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DepositPlan"
    delete:
      operationId: deletePlan
      tags: [admin]
      summary: Delete a deposit plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Plan deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/tasks:
    get:
      operationId: getAdminTasks
      tags: [admin]
      summary: Get all tasks (admin)
      responses:
        "200":
          description: Tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [admin]
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskBody"
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /admin/tasks/{id}:
    patch:
      operationId: updateTask
      tags: [admin]
      summary: Update a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskBody"
      responses:
        "200":
          description: Task updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [admin]
      summary: Delete a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /admin/withdrawals/{id}/approve:
    patch:
      operationId: approveWithdrawal
      tags: [admin]
      summary: Approve withdrawal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/withdrawals/{id}/reject:
    patch:
      operationId: rejectWithdrawal
      tags: [admin]
      summary: Reject withdrawal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectWithdrawalBody"
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/trade/settings:
    get:
      operationId: getAdminTradeSettings
      tags: [admin]
      summary: Get trade settings (admin)
      responses:
        "200":
          description: Trade settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradeSettings"
    patch:
      operationId: updateTradeDirection
      tags: [admin]
      summary: Update trade direction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTradeDirectionBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradeSettings"
  /admin/announcements:
    get:
      operationId: getAdminAnnouncements
      tags: [admin]
      summary: Get all announcements (admin)
      responses:
        "200":
          description: Announcements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Announcement"
    post:
      operationId: createAnnouncement
      tags: [admin]
      summary: Create announcement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAnnouncementBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
  /admin/announcements/{id}:
    patch:
      operationId: updateAnnouncement
      tags: [admin]
      summary: Update announcement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAnnouncementBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
    delete:
      operationId: deleteAnnouncement
      tags: [admin]
      summary: Delete announcement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/inbox/send:
    post:
      operationId: sendInboxMessage
      tags: [admin]
      summary: Send inbox message to user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendInboxMessageBody"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
...
    ReferralEntry:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        level:
          type: integer
        bonusAmount:
          type: number
        joinedAt:
          type: string
      required: [id, name, level, bonusAmount, joinedAt]
...
    DepositPlan:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        minAmount:
          type: number
        maxAmount:
          type: ["number", "null"]
        dailyRate:
          type: number
        dailyEarning:
          type: number
        durationDays:
          type: integer
        bonusPercent:
          type: number
        isActive:
          type: boolean
        description:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, name, minAmount, dailyRate, dailyEarning, durationDays, bonusPercent, isActive, createdAt]
    CreatePlanBody:
      type: object
      properties:
        name:
          type: string
        minAmount:
          type: number
        maxAmount:
          type: ["number", "null"]
        dailyRate:
          type: number
        durationDays:
          type: integer
        bonusPercent:
          type: number
        isActive:
          type: boolean
        description:
          type: ["string", "null"]
      required: [name, minAmount, dailyRate, durationDays, bonusPercent, isActive]
    Deposit:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        planId:
          type: integer
        planName:
          type: string
        amount:
          type: number
        bonusAmount:
          type: number
        dailyEarning:
          type: number
        status:
          type: string
        autoRenew:
          type: boolean
        startsAt:
          type: ["string", "null"]
        endsAt:
          type: ["string", "null"]
        lastEarningAt:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, userId, planId, planName, amount, bonusAmount, dailyEarning, status, autoRenew, createdAt]
...
    Earning:
      type: object
      properties:
        id:
          type: integer
        amount:
          type: number
        type:
          type: string
        description:
          type: string
        createdAt:
          type: string
      required: [id, amount, type, description, createdAt]
...
    Task:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        reward:
          type: number
        link:
          type: ["string", "null"]
        isActive:
          type: boolean
        isCompleted:
          type: boolean
        createdAt:
          type: string
      required: [id, title, description, reward, isActive, isCompleted, createdAt]
    CreateTaskBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        reward:
          type: number
        link:
          type: ["string", "null"]
        isActive:
          type: boolean
      required: [title, description, reward, isActive]
    Withdrawal:
      type: object
      properties:
        id:
          type: integer
        amount:
          type: number
        phone:
          type: string
        status:
          type: string
        adminNote:
          type: ["string", "null"]
        requestedAt:
          type: string
        processedAt:
          type: ["string", "null"]
      required: [id, amount, phone, status, requestedAt]
...
    TradeSettings:
      type: object
      properties:
        direction:
          type: string
        updatedAt:
          type: string
      required: [direction, updatedAt]
...
    UpdateTradeDirectionBody:
      type: object
      properties:
        direction:
          type: string
      required: [direction]
    Announcement:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
      required: [id, title, content, isActive, createdAt]
    CreateAnnouncementBody:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        isActive:
          type: boolean
      required: [title, content, isActive]
...
    SendInboxMessageBody:
      type: object
      properties:
        userId:
          type: ["integer", "null"]
        title:
          type: string
        content:
          type: string
      required: [title, content]
...
    AdminUser:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        balance:
          type: number
        totalEarned:
          type: number
        totalDeposited:
          type: number
        vipLevel:
          type: string
        referralCode:
          type: string
        referralCount:
          type: integer
        isSuspended:
          type: boolean
        createdAt:
          type: string
      required: [id, name, email, phone, balance, totalEarned, totalDeposited, vipLevel, referralCode, referralCount, isSuspended, createdAt]
    AdminUserDetail:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/AdminUser"
        deposits:
          type: array
          items:
            $ref: "#/components/schemas/Deposit"
        withdrawals:
          type: array
          items:
            $ref: "#/components/schemas/Withdrawal"
        earnings:
          type: array
          items:
            $ref: "#/components/schemas/Earning"
        referrals:
          type: array
          items:
            $ref: "#/components/schemas/ReferralEntry"
      required: [user, deposits, withdrawals, earnings, referrals]
...
    RejectWithdrawalBody:
      type: object
      properties:
        reason:
          type: string
      required: [reason]
...